requirements/base.txt
Django==3.2.*
django-environ
mysqlclient